Create nmt_src
The code was written in python language and was trained on a system that had an i5 7th gen processor.  The RAM of the system was 8GB. The entire project was implemented on the Windows operating system.
The total training time period was about 5 hours and the testing took 2 hours.